try to fix 64 bit / 32 bit problems

+++ src/eins.c
@@ -1,4 +1,4 @@
-/* 
+/*
  * eins - A tool for benchmarking networks.
  * Copyright (C) 2006  Hynek Schlawack <hs+eins@ox.cx>
  *
@@ -49,7 +49,7 @@
     "\n"\
     "Server: eins -s -t type\n" \
     "\t-s: Start as server"
-               
+
 #define DEF_OPTS "st:i:nu:b:q"
 
 static void
@@ -74,7 +74,7 @@ build_optstr(const net_mod *m[])
     for (int i = 0; m[i]; i++) {
         len += strlen(m[i]->opts);
     }
-    
+
     char *str = safe_alloc(len + 1);
     strcpy(str, DEF_OPTS);
 
@@ -101,7 +101,7 @@ parse_args(int argc, char * argv[], mod_args *ma, prefs *p)
         case 's':
             ma->mode = EINS_SERVER;
             break;
-	    
+
             // Mods
         case 't':
             for (int i = 0; Modules[i]; i++) {
@@ -155,7 +155,7 @@ parse_args(int argc, char * argv[], mod_args *ma, prefs *p)
     if ((ma->mode == EINS_CLIENT) && ((argc != (optind + 2)))) {
         usage_and_die();
     }
-  
+
     if (ma->mode == EINS_CLIENT) {
         ma->target = argv[optind];
         ma->size = atoi(argv[optind + 1]);
@@ -234,11 +234,11 @@ main(int argc, char **argv)
 	    double bw = ma.size / med;
 	    //bw *= (double) 15625 / 16384;
 
-            printf("%15d%16f%16f%16f\n", ma.size, med, bw, sqrt(var));
+            printf("%15d%16f%16f%16f\n", (int)ma.size, med, bw, sqrt(var));
         }
 
         nm->cleanup();
-	
+
         free(alltime);
         free(ma.payload);
