+++ src/mod_udp.c
@@ -1,4 +1,4 @@
-/* 
+/*
  * eins - A tool for measuring network-bandwidths and -latencies.
  * Copyright (C) 2006  Hynek Schlawack <hs+eins@ox.cx>
  *
@@ -47,8 +47,8 @@ static struct udp_prefs {
 typedef struct {
     handshake h;
     ip_handshake ip;
-    
-    size_t frag_size;
+
+    int32_t frag_size;
 } udp_handshake;
 
 
@@ -66,8 +66,8 @@ udp_handle_arg(char opt, char *arg)
     case 'F':
         Prefs.frag_size = atoi(arg);
         return true;
-	
-    default:       
+
+    default:
         return ip_handle_arg((ip_prefs *) &Prefs, opt, arg);
     }
 }
@@ -163,7 +163,7 @@ udp_serve(mod_args *ma)
             sin->sin_addr.s_addr = htonl(INADDR_ANY);
         }
         sin->sin_port = htons(atoi(Prefs.ip.port));
-  
+
         servaddr = (struct sockaddr *) sin;
         sock_size = sizeof(struct sockaddr_in);
     }
@@ -177,9 +177,9 @@ udp_serve(mod_args *ma)
     if (bind(sd, servaddr, sock_size) == -1) {
         XLE("Server: bind");
     }
-	
+
     struct sockaddr *from = safe_alloc(sock_size);
-    while (1) {		
+    while (1) {
         char *data;
 
         memset(from, 0, sock_size);
@@ -201,7 +201,7 @@ udp_serve(mod_args *ma)
                 LE("Server: sendto");
                 return false;
             }
-	    
+
         } else {
             response = 0;
             if (sendto(sd, &response, sizeof(response), 0, from, sock_size) == -1) {
@@ -213,7 +213,7 @@ udp_serve(mod_args *ma)
 
             continue;
         }
-	
+
         // Measure-loop
         for (size_t i = 0; i < uh.h.tries; i++) {
             size_t bytes;
@@ -230,8 +230,8 @@ udp_serve(mod_args *ma)
             }
 
             for (bytes = 0; bytes < uh.h.size; bytes += rc) {
-                rc = sendto(sd, data + bytes, 
-                            (bytes + uh.frag_size) > uh.h.size ? uh.h.size - bytes : uh.frag_size, 
+                rc = sendto(sd, data + bytes,
+                            (bytes + uh.frag_size) > uh.h.size ? uh.h.size - bytes : uh.frag_size,
                             0, from, sock_size);
                 if (rc == -1) {
                     LE("Server: sendto");
@@ -239,7 +239,7 @@ udp_serve(mod_args *ma)
                 }
             }
         }
-	
+
         free(data);
     }
